feat(spikesorting v2): content fingerprint + per-recording artifact lock
Add _recording_fingerprint.py: a DB-light, readback-from-disk content
fingerprint for the preprocessed recording cache. It hashes the persisted
ElectricalSeries -- traces (rounded, little-endian), timestamps (rounded),
persisted-region probe geometry, and scaling/shape metadata -- into a
component dict whose combined_hash is the recording's representation-blind
scientific identity. Unlike the whole-file NwbfileHasher digest, it is
reproducible across byte-different rebuilds, so one value can later drive
recompute matched, delete authority, and rebuild reconciliation.

Design choices baked in and documented in the module:
- opens the known abs path directly, bypassing the checksum-validating
  AnalysisNwbfile.get_abs_path (that stale checksum is the bug being fixed);
- geometry from the persisted electrodes region, with get_channel_locations
  kept only as a parity check (geometry_component_hash is shared so both
  sources hash equal);
- lineage (raw source_object_id) is deliberately NOT hashed -- it is
  provenance, not content; the traces component already discriminates source,
  and content_hash is only ever compared within one recording_id.

Also add recording_artifact_lock(recording_id), mirroring
analyzer_curation_lock on the shared analyzer/lock root, to serialize
rebuild vs reclamation of the same recording (used in later sub-units).

Make hash_recording_traces serialize explicit little-endian (no hash change
on x86/ARM) so the fingerprint is byte-stable across architectures.

Tests (DB-free): fingerprint is deterministic; discriminates perturbed
traces/timestamps/gain/channel-order/geometry while absorbing sub-rounding
noise; persisted-region geometry matches get_channel_locations (parity).

+++ src/spyglass/spikesorting/v2/_recompute.py
@@ -70,8 +70,13 @@ def hash_recording_traces(
                 ),
                 dtype=np.float64,
             )
+            # Serialize explicit little-endian so the digest is byte-stable
+            # across architectures (the content-fingerprint contract). traces
+            # is float64 above; ``<f8`` is a no-op on x86/ARM but defensive.
             digest.update(
-                np.ascontiguousarray(np.round(traces, rounding)).tobytes()
+                np.ascontiguousarray(
+                    np.round(traces, rounding).astype("<f8")
+                ).tobytes()
             )
         hashes[f"segment_{segment}"] = digest.hexdigest()
     return hashes
